update router to single index file oncemore to avoid circular dependencies and add more documentation.
